# Optimize method call performance