This would improve the performance for method calls a bit:

Microbenchmark:
```python
s = 0
for i in range(1000000):
    s += 1
```
Result:
```
Benchmark #1: ./rustpython-master test.py
  Time (mean ± σ):     870.1 ms ±   8.1 ms    [User: 867.4 ms, System: 2.5 ms]
  Range (min … max):   858.6 ms … 884.0 ms    10 runs
 
Benchmark #2: ./rustpython-opt test.py
  Time (mean ± σ):     857.7 ms ±   7.1 ms    [User: 853.8 ms, System: 3.3 ms]
  Range (min … max):   841.8 ms … 866.0 ms    10 runs
 
Summary
  './rustpython-opt test.py' ran
    1.01 ± 0.01 times faster than './rustpython-master test.py'

------

Benchmark #1: ./rustpython-master-nothreading test.py
  Time (mean ± σ):     520.8 ms ±   6.6 ms    [User: 517.9 ms, System: 2.7 ms]
  Range (min … max):   512.6 ms … 530.6 ms    10 runs
 
Benchmark #2: ./rustpython-opt-nothreading test.py
  Time (mean ± σ):     486.7 ms ±  25.9 ms    [User: 483.7 ms, System: 2.9 ms]
  Range (min … max):   464.2 ms … 551.0 ms    10 runs
 
Summary
  './rustpython-opt-nothreading test.py' ran
    1.07 ± 0.06 times faster than './rustpython-master-nothreading test.py'
```

pystone benchmark:
```
Benchmark #1: ./rustpython-master benches/benchmarks/pystone.py
  Time (mean ± σ):      2.873 s ±  0.012 s    [User: 2.869 s, System: 0.002 s]
  Range (min … max):    2.860 s …  2.902 s    10 runs
 
Benchmark #2: ./rustpython-opt benches/benchmarks/pystone.py
  Time (mean ± σ):      2.838 s ±  0.011 s    [User: 2.835 s, System: 0.002 s]
  Range (min … max):    2.825 s …  2.859 s    10 runs
 
Summary
  './rustpython-opt benches/benchmarks/pystone.py' ran
    1.01 ± 0.01 times faster than './rustpython-master benches/benchmarks/pystone.py'

------

Benchmark #1: ./rustpython-master-nothreading benches/benchmarks/pystone.py
  Time (mean ± σ):      1.879 s ±  0.023 s    [User: 1.875 s, System: 0.003 s]
  Range (min … max):    1.846 s …  1.923 s    10 runs
 
Benchmark #2: ./rustpython-opt-nothreading benches/benchmarks/pystone.py
  Time (mean ± σ):      1.845 s ±  0.025 s    [User: 1.840 s, System: 0.003 s]
  Range (min … max):    1.807 s …  1.876 s    10 runs
 
Summary
  './rustpython-opt-nothreading benches/benchmarks/pystone.py' ran
    1.02 ± 0.02 times faster than './rustpython-master-nothreading benches/benchmarks/pystone.py'
```