@coolreader18 2021-03-06
This looks good, though it would also be cool if we implemented the logic in `_PyObject_GetMethod`, to avoid the need for creating a `method` object altogether