@pca006132 2021-03-08
> This looks good, though it would also be cool if we implemented the logic in `_PyObject_GetMethod`, to avoid the need for creating a `method` object altogether

Yeah, avoiding the creation of a `method` object would be better, but I may not be able to implement that this month.

Maybe someone else can have a look at that, or wait for me to revisit this next month? :)